requirements.txt
streamlit
numpy
opencv-python
Pillow
tensorflow
scikit-image
